Molecular formula parser extracted.
