+++ utilities/analyseGaussLog.py
@@ -1,7 +1,7 @@
 """Module containing parsing functions for G16 log files."""
 
 import numpy as np
-from .utils import readlines_reverse
+from .utils import readlines_reverse, parse_mol_formula
 
 
 class GaussianLog():
@@ -127,7 +127,7 @@ class GaussianLog():
                         if 'natoms' in line.lower():
                             self.atom_number = int(line.split()[1])
                             break
-                self._pull_atom_ids()
+                self.atom_ids = self._pull_atom_ids()
 
                 print(
                     'Normal termination output not present, '
@@ -668,60 +668,3 @@ class GaussianLog():
             for line in input:
                 if 'natoms' in line.lower():
                     self.atom_number = int(line.split()[1])
-
-
-def parse_mol_formula(mol_formula):
-    """
-    Parse the molecular formula to get the atoms and charge.
-
-    Parameters
-    ----------
-    mol_formula : :class:`str`
-        The molecular formula to be parsed.
-
-    Returns
-    -------
-    :class:`int`
-        The number of atoms in the molecule.
-
-    :class:`list of str`
-        The list of elements present in the molecule.
-
-    :class:`int`
-        The charge of the molecule.
-
-    """
-    atom_number = 0
-    elements = []
-
-    # Identify charge and remove it from the molecular formula.
-    if mol_formula.isalnum():
-        charge = 0
-    else:
-        charge = int(mol_formula.split('(')[1][:-2])
-        if mol_formula[-2] == '-':
-            charge *= -1
-    mol_formula = mol_formula.split('(')[0]
-
-    # Find individual elements in the molecular formula.
-    i = 0
-
-    while i < len(mol_formula)-1:
-        char = mol_formula[i]
-        while char.isdigit() is mol_formula[i+1].isdigit():
-            char += mol_formula[i+1]
-            i += 1
-        # Identify if an element or a count.
-        if char.isdigit():
-            atom_number += int(char)
-        else:
-            elements.append(char)
-        i += 1
-        # Add last entry if not already included
-        if i != len(mol_formula):
-            if mol_formula[-1].isdigit():
-                atom_number += int(mol_formula[-1])
-            else:
-                elements.append(mol_formula[-1])
-
-    return atom_number, elements, charge
