+++ utilities/utils.py
@@ -17,3 +17,60 @@ def readlines_reverse(filename):
                 line += next_char
             position -= 1
         yield line[::-1]
+
+
+def parse_mol_formula(mol_formula):
+    """
+    Parse the molecular formula to get the atoms and charge.
+
+    Parameters
+    ----------
+    mol_formula : :class:`str`
+        The molecular formula to be parsed.
+
+    Returns
+    -------
+    :class:`int`
+        The number of atoms in the molecule.
+
+    :class:`list of str`
+        The list of elements present in the molecule.
+
+    :class:`int`
+        The charge of the molecule.
+
+    """
+    atom_number = 0
+    elements = []
+
+    # Identify charge and remove it from the molecular formula.
+    if mol_formula.isalnum():
+        charge = 0
+    else:
+        charge = int(mol_formula.split('(')[1][:-2])
+        if mol_formula[-2] == '-':
+            charge *= -1
+    mol_formula = mol_formula.split('(')[0]
+
+    # Find individual elements in the molecular formula.
+    i = 0
+
+    while i < len(mol_formula)-1:
+        char = mol_formula[i]
+        while char.isdigit() is mol_formula[i+1].isdigit():
+            char += mol_formula[i+1]
+            i += 1
+        # Identify if an element or a count.
+        if char.isdigit():
+            atom_number += int(char)
+        else:
+            elements.append(char)
+        i += 1
+        # Add last entry if not already included
+        if i != len(mol_formula):
+            if mol_formula[-1].isdigit():
+                atom_number += int(mol_formula[-1])
+            else:
+                elements.append(mol_formula[-1])
+
+    return atom_number, elements, charge
